@@ -9,14 +9,15 @@
     <description>Spring Data Elasticsearch Extension</description>
     <properties>
         <!-- Build configuration -->
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
 
         <!-- Dependencies versions -->
-        <spring.elasticsearch.version>4.4.6</spring.elasticsearch.version>
-        <spring.version>5.3.24</spring.version>
+        <spring.version>6.0.0</spring.version>
+        <spring.elasticsearch.version>5.0.0</spring.elasticsearch.version>
+        <elasticsearch-rhlc.version>7.17.8</elasticsearch-rhlc.version>
 
         <!-- test Dependencies versions -->
         <junit5.version>5.8.2</junit5.version>
