@@ -11,13 +11,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.dao.DataAccessException;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.elasticsearch.annotations.Document;
 import org.springframework.data.elasticsearch.core.SearchHitsIterator;
-import org.springframework.data.elasticsearch.core.convert.ElasticsearchConverter;
-import org.springframework.data.elasticsearch.core.convert.MappingElasticsearchConverter;
-import org.springframework.data.elasticsearch.core.mapping.SimpleElasticsearchMappingContext;
 import org.springframework.data.elasticsearch.core.query.Query;
 
 import java.io.IOException;
@@ -34,7 +32,7 @@ import static org.mockito.Mockito.*;
 class ExtendedElasticsearchTemplateTest {
 
     @Document(indexName = "test")
-    static class TestEntity{
+    static class TestEntity {
         @Id
         private String id;
     }
@@ -50,15 +48,14 @@ class ExtendedElasticsearchTemplateTest {
         when(client._transport()).thenReturn(transport);
         when(transport.jsonpMapper()).thenReturn(new SimpleJsonpMapper());
 
-        SimpleElasticsearchMappingContext mappingContext = new SimpleElasticsearchMappingContext();
-        ElasticsearchConverter elasticsearchConverter = new MappingElasticsearchConverter(mappingContext);
-        extendedElasticsearchTemplate = new ExtendedElasticsearchTemplate(client, elasticsearchConverter);
+        extendedElasticsearchTemplate = new ExtendedElasticsearchTemplate(client);
     }
 
     private List<Hit<EntityAsMap>> createResultHits(int size) {
         List<Hit<EntityAsMap>> results = new ArrayList<>();
         for (int index = 0; index < size; index++) {
-            results.add(Hit.of(builder -> builder.index("testEntity").id("id").version(1L)));
+            String id = String.valueOf(index);
+            results.add(Hit.of(builder -> builder.index("testEntity").id(id).version(1L)));
         }
         return results;
     }
@@ -112,6 +109,7 @@ class ExtendedElasticsearchTemplateTest {
 
         Query query = new NativeQueryBuilder()
                 .withPageable(Pageable.ofSize(100))
+                .withMaxResults(200)
                 .build();
         // When
         SearchHitsIterator<TestEntity> iterator = extendedElasticsearchTemplate.searchForStream(query, 100, TestEntity.class);
@@ -135,9 +133,43 @@ class ExtendedElasticsearchTemplateTest {
     void shouldThrowAssertionErrorIfPageableOfQueryIsNull() {
         Query query = mock(Query.class);
 
-
         assertThatThrownBy(() -> extendedElasticsearchTemplate.searchForStream(query, 0, TestEntity.class))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("pageable of query must not be null.");
+    }
+
+    @Test
+    void shouldHandleIOExceptionAndTransformItToDataAccessException() throws IOException {
+        // Given
+        when(client.search(any(SearchRequest.class), eq(EntityAsMap.class))).thenThrow(IOException.class);
+        Query query = new NativeQueryBuilder()
+                .withPageable(Pageable.ofSize(100))
+                .build();
+
+        // When
+        assertThatThrownBy(() -> extendedElasticsearchTemplate.searchForStream(query, 0, TestEntity.class))
+                .isInstanceOf(DataAccessException.class);
+    }
+
+    @Test
+    void shouldHandleRuntimeExceptionAndReThrowIt() throws IOException {
+        // Given
+        when(client.search(any(SearchRequest.class), eq(EntityAsMap.class))).thenThrow(RuntimeException.class);
+        Query query = new NativeQueryBuilder()
+                .withPageable(Pageable.ofSize(100))
+                .build();
+
+        // When
+        assertThatThrownBy(() -> extendedElasticsearchTemplate.searchForStream(query, 0, TestEntity.class))
+                .isInstanceOf(RuntimeException.class);
+    }
+
+    @Test
+    void shouldNotCallElasticsearchSWhenClearingEmptyScrollIdList() throws IOException {
+        // When
+        extendedElasticsearchTemplate.searchScrollClear(List.of());
+
+        // Then
+        verify(client, times(0)).clearScroll(any(ClearScrollRequest.class));
     }
 }
